requirements.txt
annotated-types
asgiref
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-bootstrap-datepicker-plus
django-cors-headers
django-crispy-forms
django-email
django-environ
django-filter
django-tempus-dominus
djangorestframework
filelock
Markdown
pillow
pipenv
platformdirs
pydantic
pydantic_core
pytz
setuptools
sqlparse
typing_extensions
tzdata
virtualenv
